add tests for np enhancements, create ST for NPs etc.
