+++ systemtest/src/test/java/io/strimzi/systemtest/connect/ConnectS2IST.java
@@ -984,7 +984,7 @@ class ConnectS2IST extends AbstractST {
     }
 
     @BeforeAll
-    void setup() throws Exception {
+    void setup() {
         ResourceManager.setClassResources();
         installClusterOperator(NAMESPACE, Constants.CO_OPERATION_TIMEOUT_SHORT, Constants.RECONCILIATION_INTERVAL);
